refactor: reorganizar estrutura do projeto com novas pastas e renomeação de arquivos
