--- meu-app/frontend/App.js
@@ -1,11 +1,0 @@
-import React from "react";
-import { NavigationContainer } from "@react-navigation/native";
-import Home from "./Home";
-
-export default function App() {
-  return (
-    <NavigationContainer>
-      <Home />{" "}
-    </NavigationContainer>
-  );
-}
