+++ meu-app/frontend/app/components/infoDiabetesItems/admInsulinaItem.js
