+++ meu-app/frontend/app/components/infoDiabetesItems/tiposDiabetesItem.js
