+++ meu-app/frontend/app/components/slidesOnboard.js
