+++ meu-app/frontend/app/components/tabBottomIcons/EmergenciaTab.js
