+++ meu-app/frontend/app/components/tabBottomIcons/HomeTab.js
