+++ meu-app/frontend/app/components/tabBottomIcons/MedicacaoTab.js
