+++ meu-app/frontend/app/components/tabBottomIcons/RelatoriosTab.js
