+++ meu-app/frontend/app/components/tabTopIcons/TabLabelHistorico.js
@@ -6,13 +6,13 @@ export default function TabLabelSaude({ focused }) {
   return (
     <View style={styles.tabLabelContainer}>
       <Image
-        source={require("../assets/images/saude-on.png")}
+        source={require("../../assets/images/historico.png")}
         style={[styles.tabImage, focused && { tintColor: "#6179FA" }]}
       />
       <Text
         style={[styles.tabText, focused ? styles.colorFocused : styles.tabText]}
       >
-        Saúde
+        Histórico
       </Text>
     </View>
   );
@@ -27,7 +27,6 @@ const styles = StyleSheet.create({
   tabImage: {
     width: 20,
     height: 20,
-    tintColor: "#7B7A78",
   },
   tabText: {
     fontSize: 16,
