+++ meu-app/frontend/app/components/tabTopIcons/TabLabelSaude.js
@@ -6,13 +6,13 @@ export default function TabLabelSaude({ focused }) {
   return (
     <View style={styles.tabLabelContainer}>
       <Image
-        source={require("../assets/images/historico.png")}
+        source={require("../../assets/images/saude-on.png")}
         style={[styles.tabImage, focused && { tintColor: "#6179FA" }]}
       />
       <Text
         style={[styles.tabText, focused ? styles.colorFocused : styles.tabText]}
       >
-        Histórico
+        Saúde
       </Text>
     </View>
   );
@@ -27,6 +27,7 @@ const styles = StyleSheet.create({
   tabImage: {
     width: 20,
     height: 20,
+    tintColor: "#7B7A78",
   },
   tabText: {
     fontSize: 16,
