+++ meu-app/frontend/app/screens/Feedbacks/checkSucess.jsx
@@ -9,26 +9,30 @@ import {
 import React from "react";
 import { useNavigation } from "@react-navigation/native";
 
-export default function checkFailed() {
+export default function checkSucess() {
   const navigation = useNavigation();
   return (
     <KeyboardAvoidingView style={styles.container}>
       <View>
         <Image
           style={styles.image}
-          source={require("../assets/images/failed.webp")}
+          source={require("../../assets/images/sucess.webp")}
         />
       </View>
 
       <View style={styles.containerText}>
-        <Text style={styles.textContainer}>Cadastro não realizado!</Text>
+        <Text style={styles.textContainer}>
+          Cadastro realizado com sucesso!
+        </Text>
       </View>
 
-      <TouchableOpacity
-        style={styles.btn}
-        onPress={() => navigation.navigate("screens/Auth/telaCadastro")}
-      >
-        <Text style={styles.textBtn}>Tentar novamente</Text>
+      <TouchableOpacity style={styles.btn}>
+        <Text
+          style={styles.textBtn}
+          onPress={() => navigation.navigate("screens/login")}
+        >
+          Ir para login
+        </Text>
       </TouchableOpacity>
     </KeyboardAvoidingView>
   );
