+++ meu-app/frontend/app/screens/Feedbacks/recoverAccountEmail.jsx
@@ -1,15 +1,6 @@
 import React, { useEffect } from "react";
-import {
-  Text,
-  View,
-  Image,
-  TouchableOpacity,
-  Button,
-  StyleSheet,
-} from "react-native";
+import { Text, View, Image, TouchableOpacity, StyleSheet } from "react-native";
 import { useNavigation, useRouter } from "expo-router";
-import { useFonts, Lato_400Regular } from "@expo-google-fonts/lato";
-import { Urbanist_700Bold } from "@expo-google-fonts/urbanist";
 
 export default function recoverAccountEmail() {
   const navigation = useNavigation();
@@ -20,18 +11,13 @@ export default function recoverAccountEmail() {
     navigation.setOptions({ headerShown: false });
   }, [navigation]);
 
-  const [fonteLoaded] = useFonts({
-    Lato_400Regular,
-    Urbanist_700Bold,
-  });
-
   return (
     <View style={{ flex: 1 }}>
       <View style={{ marginTop: 46, paddingHorizontal: 20 }}>
         <TouchableOpacity onPress={() => navigation.goBack()}>
           <Image
             style={{ width: 24, height: 24 }}
-            source={require("../assets/images/backIcon.png")}
+            source={require("../../assets/images/backIcon.png")}
           />
         </TouchableOpacity>
       </View>
@@ -40,16 +26,18 @@ export default function recoverAccountEmail() {
         <View>
           <Image
             style={styles.image}
-            source={require("../assets/images/undraw_confirmed_re_sef71.png")}
+            source={require("../../assets/images/undraw_mail_sent_re_0ofv1.png")}
           />
         </View>
 
         <View>
-          <Text style={styles.text}>Senha recuperada com sucesso!</Text>
+          <Text style={styles.text}>
+            Verifique seu e-mail e abra o link que enviamos para continuar!
+          </Text>
         </View>
 
-        <TouchableOpacity style={styles.btn} onPress={router.push("./login")}>
-          <Text style={styles.textBtn}>Login</Text>
+        <TouchableOpacity style={styles.btn} onPress={router.push("./welcome")}>
+          <Text style={styles.textBtn}>OK</Text>
         </TouchableOpacity>
       </View>
     </View>
@@ -64,12 +52,13 @@ const styles = StyleSheet.create({
   },
 
   text: {
-    fontFamily: "Urbanist_700Bold",
+    fontFamily: "Lato_400Regular",
     fontSize: 20,
     fontStyle: "normal",
-    lineHeight: 22,
+    lineHeight: 27,
     textAlign: "center",
-    color: "#373737",
+    color: "#282828",
+    paddingHorizontal: 30,
   },
   btn: {
     backgroundColor: "#2F39D3",
