+++ meu-app/frontend/app/screens/HomeBottomBar/emergencia.jsx
@@ -3,7 +3,7 @@ import { Text, View } from "react-native";
 function Emergencia() {
   return (
     <View>
-      <Text>teste</Text>
+      <Text>Emergência</Text>
     </View>
   );
 }
