+++ meu-app/frontend/app/screens/HomeBottomBar/medicacao.jsx
@@ -1,11 +1,11 @@
 import { Text, View } from "react-native";
 
-function Relatorios() {
+function Medicacao() {
   return (
     <View>
-      <Text>teste</Text>
+      <Text>Medicação</Text>
     </View>
   );
 }
 
-export default Relatorios;
+export default Medicacao;
